test(grpc): isolate grpc protobuf package into separate module
