+++ test/apps/grpcserver/go.mod
@@ -2,9 +2,11 @@ module github.com/open-telemetry/opentelemetry-go-compile-instrumentation/test/a
 
 go 1.25.0
 
+replace github.com/open-telemetry/opentelemetry-go-compile-instrumentation/test/shared/grpcpb => ../../shared/grpcpb
+
 require (
+	github.com/open-telemetry/opentelemetry-go-compile-instrumentation/test/shared/grpcpb v0.0.0-00010101000000-000000000000
 	google.golang.org/grpc v1.80.0
-	google.golang.org/protobuf v1.36.11
 )
 
 require (
@@ -14,4 +16,5 @@ require (
 	golang.org/x/sys v0.41.0 // indirect
 	golang.org/x/text v0.33.0 // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20260120221211-b8f7ae30c516 // indirect
+	google.golang.org/protobuf v1.36.11 // indirect
 )
