+++ test/shared/grpcpb/go.mod
@@ -1,0 +1,15 @@
+module github.com/open-telemetry/opentelemetry-go-compile-instrumentation/test/shared/grpcpb
+
+go 1.25.0
+
+require (
+	google.golang.org/grpc v1.80.0
+	google.golang.org/protobuf v1.36.11
+)
+
+require (
+	golang.org/x/net v0.49.0 // indirect
+	golang.org/x/sys v0.40.0 // indirect
+	golang.org/x/text v0.33.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20260120221211-b8f7ae30c516 // indirect
+)
