+++ test/shared/grpcpb/greeter.proto
@@ -5,7 +5,7 @@ syntax = "proto3";
 
 package greeter;
 
-option go_package = "github.com/open-telemetry/opentelemetry-go-compile-instrumentation/test/apps/grpcserver/pb";
+option go_package = "github.com/open-telemetry/opentelemetry-go-compile-instrumentation/test/shared/grpcpb/pb";
 
 service Greeter {
   rpc SayHello (HelloRequest) returns (HelloReply) {}
